Update scripts to watch script.js. Convert uint's to uint256's.

+++ contracts/Airdrop.sol
@@ -14,15 +14,15 @@ contract Airdrop is AragonApp {
     }
 
     /// Events
-    event Start(uint id);
-    event Award(uint id, address recipient, uint amount);
+    event Start(uint256 id);
+    event Award(uint256 id, address recipient, uint256 amount);
 
     /// State
     Agent public agent;
     ICycleManager public cycleManager;
     address public sctAddress;
 
-    mapping(uint => Airdrop) public airdrops;
+    mapping(uint256 => Airdrop) public airdrops;
     uint256 public airdropsCount;
     uint256 public lastRewardCycle;
 
@@ -55,7 +55,7 @@ contract Airdrop is AragonApp {
         _start(_root, _dataURI);
     }
 
-    function _start(bytes32 _root, string _dataURI) internal returns(uint id) {
+    function _start(bytes32 _root, string _dataURI) internal returns(uint256 id) {
         require(cycleManager.currentCycle() > lastRewardCycle, ERROR_CYCLE_NOT_ENDED);
         lastRewardCycle = cycleManager.currentCycle();
 
@@ -71,7 +71,7 @@ contract Airdrop is AragonApp {
      * @param _amount The token amount
      * @param _proof Merkle proof to correspond to data supplied
      */
-    function award(uint _id, address _recipient, uint256 _amount, bytes32[] _proof) public {
+    function award(uint256 _id, address _recipient, uint256 _amount, bytes32[] _proof) public {
         Airdrop storage airdrop = airdrops[_id];
 
         bytes32 hash = keccak256(abi.encodePacked(_recipient, _amount));
@@ -96,12 +96,12 @@ contract Airdrop is AragonApp {
      */
     function awardFromMany(uint[] _ids, address _recipient, uint[] _amounts, bytes _proofs, uint[] _proofLengths) public {
 
-        uint totalAmount;
+        uint256 totalAmount;
 
-        uint marker = 32;
+        uint256 marker = 32;
 
-        for (uint i = 0; i < _ids.length; i++) {
-            uint id = _ids[i];
+        for (uint256 i = 0; i < _ids.length; i++) {
+            uint256 id = _ids[i];
 
             bytes32[] memory proof = extractProof(_proofs, marker, _proofLengths[i]);
             marker += _proofLengths[i]*32;
@@ -129,11 +129,11 @@ contract Airdrop is AragonApp {
      * @param _proofs Merkle proofs
      * @param _proofLengths Merkle proof lengths
      */
-    function awardToMany(uint _id, address[] _recipients, uint[] _amounts, bytes _proofs, uint[] _proofLengths) public {
+    function awardToMany(uint256 _id, address[] _recipients, uint[] _amounts, bytes _proofs, uint[] _proofLengths) public {
 
-        uint marker = 32;
+        uint256 marker = 32;
 
-        for (uint i = 0; i < _recipients.length; i++) {
+        for (uint256 i = 0; i < _recipients.length; i++) {
             address recipient = _recipients[i];
 
             if( airdrops[_id].awarded[recipient] )
@@ -155,13 +155,13 @@ contract Airdrop is AragonApp {
 
     }
 
-    function extractProof(bytes _proofs, uint _marker, uint proofLength) public pure returns (bytes32[] proof) {
+    function extractProof(bytes _proofs, uint256 _marker, uint256 proofLength) public pure returns (bytes32[] proof) {
 
         proof = new bytes32[](proofLength);
 
         bytes32 el;
 
-        for (uint j = 0; j < proofLength; j++) {
+        for (uint256 j = 0; j < proofLength; j++) {
             assembly {
                 el := mload(add(_proofs, _marker))
             }
@@ -173,7 +173,7 @@ contract Airdrop is AragonApp {
 
     function validate(bytes32 root, bytes32[] proof, bytes32 hash) public pure returns (bool) {
 
-        for (uint i = 0; i < proof.length; i++) {
+        for (uint256 i = 0; i < proof.length; i++) {
             if (hash < proof[i]) {
                 hash = keccak256(abi.encodePacked(hash, proof[i]));
             } else {
@@ -189,7 +189,7 @@ contract Airdrop is AragonApp {
      * @param _id Airdrop id
      * @param _recipient Recipient to check
      */
-    function awarded(uint _id, address _recipient) public view returns(bool) {
+    function awarded(uint256 _id, address _recipient) public view returns(bool) {
         return airdrops[_id].awarded[_recipient];
     }
 }
